# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SE-Textile Backend API
  description: Dynamically generated OpenAPI spec built from registered Flask routes.
  version: 2.0.0
servers:
- url: http://127.0.0.1:5001
  description: Local development server
security:
- bearerAuth: []
paths:
...
  /api/v1/customer/shops/nearby-shops:
    get:
      tags:
      - shop_explorer
      summary: Find nearby textile shops using MapMyIndia API.
      description: Find nearby textile shops using MapMyIndia API.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      parameters:
      - name: lat
        in: query
        required: false
        schema:
          type: string
      - name: lon
        in: query
        required: false
        schema:
          type: string
  /api/v1/customer/shops/search:
    get:
      tags:
      - shop_explorer
      summary: Search both shops and fabrics dynamically.
      description: Search both shops and fabrics dynamically.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
  /api/v1/customer/shops/shop/<int:shop_id>:
    get:
      tags:
      - shop_explorer
      summary: Fetch detailed shop info with all its listed fabrics.
      description: Fetch detailed shop info with all its listed fabrics.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      parameters:
      - name: shop_id
        in: path
        required: true
        schema:
          type: string
  /api/v1/customer/shops/shops:
    get:
      tags:
      - shop_explorer
      summary: Fetch all registered shops with key summary details.
      description: Fetch all registered shops with key summary details.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GenericSuccess:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
    GenericError:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        error:
          type: string
...